Move condition editor into its own file
