+++ src/panels/lovelace/editor/config-elements/hui-conditional-card-editor.ts
@@ -1,23 +1,15 @@
 import "@material/mwc-tab-bar/mwc-tab-bar";
 import "@material/mwc-tab/mwc-tab";
 import type { MDCTabBarActivatedEvent } from "@material/tab-bar";
-import {
-  mdiCodeBraces,
-  mdiContentCopy,
-  mdiListBoxOutline,
-  mdiPlus,
-} from "@mdi/js";
+import { mdiCodeBraces, mdiContentCopy, mdiListBoxOutline } from "@mdi/js";
 import deepClone from "deep-clone-simple";
 import { CSSResultGroup, LitElement, css, html, nothing } from "lit";
 import { customElement, property, query, state } from "lit/decorators";
 import { any, array, assert, assign, object, optional } from "superstruct";
 import { storage } from "../../../../common/decorators/storage";
 import { HASSDomEvent, fireEvent } from "../../../../common/dom/fire_event";
-import { stopPropagation } from "../../../../common/dom/stop_propagation";
 import "../../../../components/ha-button";
 import "../../../../components/ha-list-item";
-import "../../../../components/ha-menu-button";
-import type { HaSelect } from "../../../../components/ha-select";
 import "../../../../components/ha-svg-icon";
 import type {
   LovelaceCardConfig,
@@ -25,26 +17,16 @@ import type {
 } from "../../../../data/lovelace";
 import type { HomeAssistant } from "../../../../types";
 import type { ConditionalCardConfig } from "../../cards/types";
-import { ICON_CONDITION } from "../../common/icon-condition";
-import { Condition } from "../../common/validate-condition";
 import type { LovelaceCardEditor } from "../../types";
 import "../card-editor/hui-card-element-editor";
 import type { HuiCardElementEditor } from "../card-editor/hui-card-element-editor";
 import "../card-editor/hui-card-picker";
-import "../conditions/ha-card-condition-editor";
-import { LovelaceConditionEditorConstructor } from "../conditions/types";
-import "../conditions/types/ha-card-condition-screen";
-import "../conditions/types/ha-card-condition-state";
+import "../conditions/ha-card-conditions-editor";
 import "../hui-element-editor";
 import type { ConfigChangedEvent } from "../hui-element-editor";
 import { baseLovelaceCardConfig } from "../structs/base-card-struct";
 import type { GUIModeChangedEvent } from "../types";
 import { configElementStyle } from "./config-elements-style";
-
-const UI_CONDITION = [
-  "state",
-  "screen",
-] as const satisfies readonly Condition["condition"][];
 
 const cardConfigStruct = assign(
   baseLovelaceCardConfig,
@@ -162,59 +144,12 @@ export class HuiConditionalCardEditor
             </div>
           `
         : html`
-            <div class="conditions">
-              <ha-alert alert-type="info">
-                ${this.hass!.localize(
-                  "ui.panel.lovelace.editor.card.conditional.condition_explanation"
-                )}
-              </ha-alert>
-              ${this._config.conditions.map((cond, idx) => {
-                const condition: Condition = {
-                  condition: "state",
-                  ...cond,
-                };
-                return html`
-                  <div class="condition">
-                    <ha-card-condition-editor
-                      .index=${idx}
-                      @value-changed=${this._conditionChanged}
-                      .hass=${this.hass}
-                      .condition=${condition}
-                    ></ha-card-condition-editor>
-                  </div>
-                `;
-              })}
-              <div>
-                <ha-button-menu
-                  @action=${this._addCondition}
-                  fixed
-                  @closed=${stopPropagation}
-                >
-                  <ha-button
-                    slot="trigger"
-                    outlined
-                    .label=${this.hass.localize(
-                      "ui.panel.lovelace.editor.card.conditional.add_condition"
-                    )}
-                  >
-                    <ha-svg-icon .path=${mdiPlus} slot="icon"></ha-svg-icon>
-                  </ha-button>
-                  ${UI_CONDITION.map(
-                    (condition) => html`
-                      <ha-list-item .value=${condition} graphic="icon">
-                        ${this.hass!.localize(
-                          `ui.panel.lovelace.editor.card.conditional.condition.${condition}.label`
-                        ) || condition}
-                        <ha-svg-icon
-                          slot="graphic"
-                          .path=${ICON_CONDITION[condition]}
-                        ></ha-svg-icon>
-                      </ha-list-item>
-                    `
-                  )}
-                </ha-button-menu>
-              </div>
-            </div>
+            <ha-card-conditions-editor
+              .hass=${this.hass}
+              .conditions=${this._config.conditions}
+              @value-changed=${this._conditionChanged}
+            >
+            </ha-card-conditions-editor>
           `}
     `;
   }
@@ -281,40 +216,13 @@ export class HuiConditionalCardEditor
     fireEvent(this, "config-changed", { config: this._config });
   }
 
-  private _addCondition(ev: CustomEvent): void {
-    const condition = (ev.currentTarget as HaSelect).items[ev.detail.index]
-      .value as Condition["condition"];
+  private _conditionChanged(ev: CustomEvent) {
+    ev.stopPropagation();
     if (!this._config) {
       return;
     }
-    const conditions = [...this._config.conditions];
-
-    const elClass = customElements.get(`ha-card-condition-${condition}`) as
-      | LovelaceConditionEditorConstructor
-      | undefined;
-
-    conditions.push(
-      elClass?.defaultConfig
-        ? { ...elClass.defaultConfig }
-        : { condition: condition }
-    );
+    const conditions = ev.detail.value;
     this._config = { ...this._config, conditions };
-    fireEvent(this, "config-changed", { config: this._config });
-  }
-
-  private _conditionChanged(ev: CustomEvent) {
-    ev.stopPropagation();
-    const conditions = [...this._config!.conditions];
-    const newValue = ev.detail.value;
-    const index = (ev.target as any).index;
-
-    if (newValue === null) {
-      conditions.splice(index, 1);
-    } else {
-      conditions[index] = newValue;
-    }
-
-    this._config = { ...this._config!, conditions };
     fireEvent(this, "config-changed", { config: this._config });
   }
 
@@ -324,19 +232,6 @@ export class HuiConditionalCardEditor
       css`
         mwc-tab-bar {
           border-bottom: 1px solid var(--divider-color);
-        }
-        .conditions {
-          margin-top: 8px;
-        }
-        .condition {
-          margin-top: 8px;
-          border: 1px solid var(--divider-color);
-        }
-        .condition .content {
-          padding: 12px;
-        }
-        ha-button-menu {
-          margin-top: 12px;
         }
         .card {
           margin-top: 8px;
